# Lazy load profile images for past prime ministers