@Nooshu 2019-08-20
Excellent work Alex. This simple change makes quite a difference to the user-centric performance metrics for this page due to the large number of images:

| Metric                 | Without LL (ms) | With LL (ms) | diff (%) |
|------------------------|-----------------|--------------|----------|
| First Contentful Paint | 3271            | 2520         | -22      |
| First Meaningful Paint | 3393            | 2578         | -24      |
| Time to Interactive    | 3325            | 2840         | -15      |
| Speed Index            | 3564            | 3240         | -9       |
| Fully Loaded           | 7280            | 5821         | -20      |

[WPT comparison](https://www.webpagetest.org/video/compare.php?tests=190820_2Q_d9c0b6d6543174427b062ef38bf06c6f%2C190820_EE_777c483684b175f06ec4d929e9108c22&thumbSize=200&ival=100&end=visual) on a 3G Fast connection, Chrome version 78.0.3880.4.